+++ packages/json/src/JSONA.ts
@@ -3,7 +3,7 @@ import { JSONAError } from './JSONAError';
 
 export class JSONA {
   public static parse<T extends ObjectLiteral = ObjectLiteral>(text: string): Promise<T> {
-    return new Promise<T>((resolve: Resolve<T>, reject: Reject) => {
+    return new Promise((resolve: Resolve<T>, reject: Reject) => {
       setTimeout(() => {
         try {
           resolve(JSON.parse(text) as T);
@@ -22,7 +22,7 @@ export class JSONA {
   }
 
   public static stringify(value: ObjectLiteral): Promise<string> {
-    return new Promise<string>((resolve: Resolve<string>, reject: Reject) => {
+    return new Promise((resolve: Resolve<string>, reject: Reject) => {
       setTimeout(() => {
         try {
           resolve(JSON.stringify(value));