+++ packages/random/src/Random.ts
@@ -1,3 +1,4 @@
+import { Kind } from '@jamashita/anden-type';
 import MersenneTwister from 'mersenne-twister';
 import { RandomError } from './RandomError';
 
@@ -11,8 +12,14 @@ export class Random {
    * @param max
    */
   public static integer(min: number, max: number): number {
+    if (!Kind.isInteger(min)) {
+      throw new RandomError(`MIN IS NOT INTEGER: GIVEN: ${min}`);
+    }
+    if (!Kind.isInteger(max)) {
+      throw new RandomError(`MAX IS NOT INTEGER: GIVEN: ${max}`);
+    }
     if (min > max) {
-      throw new RandomError(`MIN IS GREATER THAN MAX: GIVEN: min = ${min}, MAX = ${max}`);
+      throw new RandomError(`MIN IS GREATER THAN MAX: GIVEN: min = ${min}, max = ${max}`);
     }
     if (min === max) {
       return min;